+++ examples/security_analysis.py
@@ -31,14 +31,21 @@ with warnings.catch_warnings():
                                  param=param,
                                  test=test
                                  )
+    multi_mix_env_pp = grid2op.make(env_name,
+                                    # ignore the protection, that are NOT simulated
+                                    # by the TimeSerie module !
+                                    param=param,
+                                    test=test
+                                    )
 
 key_env = max([el for el in multi_mix_env.keys()])
 env = multi_mix_env[key_env]
+env_pp = multi_mix_env_pp[key_env]
 
 # Run the environment on a scenario using the TimeSerie module
 security_analysis = SecurityAnalysis(env)
 security_analysis.add_all_n1_contingencies()
-a_or, voltages = security_analysis.get_flows()
+p_or, a_or, voltages = security_analysis.get_flows()
 # the 3 lines above are the only lines you need to do to perform a security analysis !
 
 computer = security_analysis.computer
@@ -64,11 +71,25 @@ for cont_id in range(env.n_line):
     sim_obs, sim_r, sim_d, sim_info = obs.simulate(env.action_space({"set_line_status": [(cont_id, -1)]}),
                                                    time_step=0)
 end_ = time.perf_counter()
-total_time = end_ - beg_
+total_time_glop_ls = end_ - beg_
+
+obs_pp = env_pp.get_obs()
+beg_ = time.perf_counter()
+for cont_id in range(env.n_line):
+    sim_obs, sim_r, sim_d, sim_info = obs_pp.simulate(env.action_space({"set_line_status": [(cont_id, -1)]}),
+                                                      time_step=0)
+end_ = time.perf_counter()
+total_time_glop_pp = end_ - beg_
+
+full_time_sa = computer.total_time() + computer.amps_computation_time()
+print()
 print("Comparison with raw grid2op timings")
-print(f"It took grid2op: {end_ - beg_:.2f}s to perform the same computation")
-print(f"This is a {(end_ - beg_) / ( computer.total_time() + computer.amps_computation_time()) :.1f} "
-      f"speed up from SecurityAnalysis over raw grid2op (using obs.simulate)")
+print(f"It took grid2op (with lightsim2grid, using obs.simulate): {total_time_glop_ls:.2f}s to perform the same computation")
+print(f"This is a {(total_time_glop_ls) / (full_time_sa) :.1f} "
+      f"speed up from SecurityAnalysis over raw grid2op (using obs.simulate and lightsim2grid)")
+print(f"It took grid2op (with pandapower, using obs.simulate): {total_time_glop_pp:.2f}s to perform the same computation")
+print(f"This is a {(total_time_glop_pp) / (full_time_sa) :.1f} "
+      f"speed up from SecurityAnalysis over raw grid2op (using obs.simulate and pandapower)")
 
 
 #### Check that the results match
@@ -77,7 +98,15 @@ for cont_id in range(env.n_line):
     sim_obs, sim_r, sim_d, sim_info = obs.simulate(action,
                                                    time_step=0)
     if not sim_d:
+        # simulation converged
+        assert np.all(np.isfinite(a_or[cont_id,:])), f"amps should not be Nan for cont {cont_id} (converged)"
         if np.max(np.abs(sim_obs.a_or - a_or[cont_id,:])) > 1e-4:
             raise RuntimeError(f"wrong amps for contingency {cont_id}")
+        if np.max(np.abs(sim_obs.p_or - p_or[cont_id,:])) > 1e-4:
+            raise RuntimeError(f"wrong active power for contingency {cont_id}")
     else:
-        assert np.all(~np.isfinite(a_or[cont_id,:])), f"amps should be Nan for cont {cont_id}"
+        # simulation diverged
+        assert np.all(~np.isfinite(a_or[cont_id,:])), f"amps should be Nan for cont {cont_id} (diverged)"
+        assert np.all(np.abs(p_or[cont_id,:]) <= 1e-6), f"active power should be 0. for cont {cont_id}"
+
+print("All results match !")