+++ examples/time_serie.py
@@ -1,4 +1,4 @@
-#!/bin/python3
+# #!/bin/python3
 
 # Copyright (c) 2020, RTE (https://www.rte-france.com)
 # See AUTHORS.txt
@@ -39,13 +39,25 @@ with warnings.catch_warnings():
                                  # by the TimeSerie module !
                                  data_feeding_kwargs={"gridvalueClass": GridStateFromFileWithForecastsWithoutMaintenance}
                                  )
+
+    multi_mix_env_pp = grid2op.make(env_name,
+                                    # ignore the protection, that are NOT simulated
+                                    # by the TimeSerie module !
+                                    param=param,
+                                    test=test,
+                                    # ignore the maintenance, that are NOT simulated
+                                    # by the TimeSerie module !
+                                    data_feeding_kwargs={"gridvalueClass": GridStateFromFileWithForecastsWithoutMaintenance}
+                                    )
 key_env = max([el for el in multi_mix_env.keys()])
 env = multi_mix_env[key_env]
+env_pp = multi_mix_env_pp[key_env]
 
 # Run the environment on a scenario using the TimeSerie module
 time_serie = TimeSerie(env)
 time_serie.compute_V(scenario_id=scenario_id)
 a_or = time_serie.compute_A()
+p_or = time_serie.compute_P()
 computer = time_serie.computer
 
 print(f"For environment: {env_name}")
@@ -70,9 +82,24 @@ done = False
 while not done:
       *_, done, info = env.step(env.action_space())
 end_ = time.perf_counter()
+total_time_glop_ls = end_ - beg_
+
+env_pp.set_id(scenario_id)
+_ = env_pp.reset()
+beg_ = time.perf_counter()
+done = False
+while not done:
+      *_, done, info = env_pp.step(env.action_space())
+end_ = time.perf_counter()
+total_time_glop_pp = end_ - beg_
+
+full_time_ts =  computer.total_time() + computer.amps_computation_time()
+print()
 print("Comparison with raw grid2op timings")
-print(f"It took grid2op: {end_ - beg_:.2f}s to perform the same computation")
-print(f"This is a {(end_ - beg_) / ( computer.total_time() + computer.amps_computation_time()) :.1f} speed up from TimeSerie over raw grid2op")
+print(f"It took grid2op (with lightsim2grid): {total_time_glop_ls:.2f}s to perform the same computation")
+print(f"This is a {(total_time_glop_ls) / (full_time_ts) :.1f} speed up from TimeSerie over raw grid2op (lightsim2grid)")
+print(f"It took grid2op (with pandapower): {total_time_glop_pp:.2f}s to perform the same computation")
+print(f"This is a {(total_time_glop_pp) / (full_time_ts) :.1f} speed up from TimeSerie over raw grid2op (pandapower)")
 
 #### Check that the results matches
 env.set_id(scenario_id)
@@ -82,5 +109,9 @@ ts = 1
 done = False
 while not done:
       obs, reward, done, info = env.step(env.action_space())
+      assert np.all(np.isfinite(a_or[ts])), f"non finite value for the amps"
+      assert np.all(np.isfinite(p_or[ts])), f"non finite value for the active power"
       assert np.max(np.abs(obs.a_or - a_or[ts])) <= 1e-4, f"wrong amps at step {ts}"
+      assert np.max(np.abs(obs.p_or - p_or[ts])) <= 1e-4, f"wrong active power at step {ts}"
       ts += 1
+print("All results match !")