+++ src/KLUSolver.cpp
@@ -8,30 +8,37 @@
 
 #include "KLUSolver.h"
 
-void KLULinearSolver::reset(){
+ErrorType KLULinearSolver::reset(){
     klu_free_symbolic(&symbolic_, &common_);
     klu_free_numeric(&numeric_, &common_);
     common_ = klu_common();
     symbolic_ = nullptr;
     numeric_ = nullptr;
+    return ErrorType::NoError;
 }
 
-bool KLULinearSolver::initialize(Eigen::SparseMatrix<real_type>&  J){
+ErrorType KLULinearSolver::initialize(Eigen::SparseMatrix<real_type>&  J){
     // default Eigen representation: column major, which is good for klu !
     // J is const here, even if it's not said in klu_analyze
     const auto n = J.cols();
     common_ = klu_common();
+    ErrorType res = ErrorType::NoError; 
     symbolic_ = klu_analyze(n, J.outerIndexPtr(), J.innerIndexPtr(), &common_);
-    numeric_ = klu_factor(J.outerIndexPtr(), J.innerIndexPtr(), J.valuePtr(), symbolic_, &common_);
-    return common_.status == KLU_OK;
+    if(common_.status != KLU_OK){
+        res = ErrorType::SolverAnalyze; 
+    }else{
+        numeric_ = klu_factor(J.outerIndexPtr(), J.innerIndexPtr(), J.valuePtr(), symbolic_, &common_);
+        if(common_.status != KLU_OK) res = ErrorType::SolverFactor; 
+    }
+    return res;
 }
 
-int KLULinearSolver::solve(Eigen::SparseMatrix<real_type>& J, RealVect & b, bool has_just_been_initialized){
+ErrorType KLULinearSolver::solve(Eigen::SparseMatrix<real_type>& J, RealVect & b, bool has_just_been_initialized){
     // solves (for x) the linear system J.x = b
     // supposes that the solver has been initialized (call klu_solver.analyze() before calling that)
     // J is const even if it does not compile if said const
     int ok;
-    int err = 0;
+    ErrorType err = ErrorType::NoError;
     bool stop = false;
     if(!has_just_been_initialized){
         // if the call to "klu_factor" has been made this iteration, there is no need
@@ -39,8 +46,8 @@ int KLULinearSolver::solve(Eigen::SparseMatrix<real_type>& J, RealVect & b, bool
         // i'm in the case where it has not
         ok = klu_refactor(J.outerIndexPtr(), J.innerIndexPtr(), J.valuePtr(), symbolic_, numeric_, &common_);
         if (ok != 1) {
-            std::cout << "\t KLU: refactor error" << std::endl;
-            err = 2;
+            // std::cout << "\t KLU: refactor error" << std::endl;
+            err = ErrorType::SolverReFactor;
             stop = true;
         }
     }
@@ -48,8 +55,8 @@ int KLULinearSolver::solve(Eigen::SparseMatrix<real_type>& J, RealVect & b, bool
         const auto n = J.cols();
         ok = klu_solve(symbolic_, numeric_, n, 1, &b(0), &common_);
         if (ok != 1) {
-            std::cout << "\t KLU: klu_solve error" << std::endl;
-            err = 3;
+            // std::cout << "\t KLU: klu_solve error" << std::endl;
+            err = ErrorType::SolverSolve;
         }
     }
     return err;