+++ src/KLUSolver.h
@@ -53,9 +53,9 @@ class KLULinearSolver
          }
 
         // public api
-        void reset();
-        bool initialize(Eigen::SparseMatrix<real_type>& J);
-        int solve(Eigen::SparseMatrix<real_type>& J, RealVect & b, bool has_just_been_inialized);
+        ErrorType reset();
+        ErrorType initialize(Eigen::SparseMatrix<real_type>& J);
+        ErrorType solve(Eigen::SparseMatrix<real_type>& J, RealVect & b, bool has_just_been_inialized);
 
     private:
         // solver initialization