+++ src/NICSLUSolver.cpp
@@ -8,7 +8,7 @@
 
 #include "NICSLUSolver.h"
 
-void NICSLULinearSolver::reset(){
+ErrorType NICSLULinearSolver::reset(){
     // free everything
     solver_.Free();
     if(ai_ != nullptr) delete [] ai_;  // created in NICSLUSolver::initialize
@@ -25,18 +25,18 @@ void NICSLULinearSolver::reset(){
             msg += "Please copy this file at the location you want to use the NICSLU solver.";
             std::cout << msg << std::endl;
         }
-        std::cout << "NICSLULinearSolver::reset() solver_.Initialize error: "  << ret << std::endl;
-        return false;
+        // std::cout << "NICSLULinearSolver::reset() solver_.Initialize error: "  << ret << std::endl;
+        return ErrorType::LicenseError;
     }
     // solver_.SetConfiguration(0, 1.); //enable timer
-    return true;
+    return ErrorType::NoError;
 }
 
-int NICSLULinearSolver::initialize(Eigen::SparseMatrix<real_type> & J){
+ErrorType NICSLULinearSolver::initialize(Eigen::SparseMatrix<real_type> & J){
     // default Eigen representation: column major, which is good for klu !
     const auto n = J.cols(); // should be equal to J_.nrows()
     int ret;
-    int err = 0; // reset error message
+    int err = ErrorType::NoError; // reset error message
     const unsigned int nnz = J.nonZeros();
 
     ai_ = new unsigned int [nnz];  // deleted in destructor and NICSLUSolver::reset
@@ -56,23 +56,23 @@ int NICSLULinearSolver::initialize(Eigen::SparseMatrix<real_type> & J){
                           ap_,
                           MATRIX_COLUMN_REAL);  // MATRIX_COLUMN_REAL, MATRIX_ROW_REAL
     if (ret < 0){
-        err = 1;
-        std::cout << "NICSLULinearSolver::initialize() solver_.Analyze error: "  << ret << std::endl;
+        err = ErrorType::SolverAnalyze;
+        // std::cout << "NICSLULinearSolver::initialize() solver_.Analyze error: "  << ret << std::endl;
         // https://github.com/chenxm1986/nicslu/blob/master/nicslu202103/include/nicslu.h for error info
         return;
     }
     // solver.FactorizeMatrix(ax, 0); //use all created threads
     ret = solver_.FactorizeMatrix(J_.valuePtr(), nb_thread_);
     if (ret < 0){
-        err = 1;
-        std::cout << "NICSLULinearSolver::initialize() solver_.FactorizeMatrix error: "  << ret << std::endl;
+        err = ErrorType::SolverFactor;
+        // std::cout << "NICSLULinearSolver::initialize() solver_.FactorizeMatrix error: "  << ret << std::endl;
         // https://github.com/chenxm1986/nicslu/blob/master/nicslu202103/include/nicslu.h for error info
         return;
     }
     return err;
 }
 
-int NICSLULinearSolver::solve(Eigen::SparseMatrix<real_type> & J, RealVect & b, bool has_just_been_inialized){
+ErrorType NICSLULinearSolver::solve(Eigen::SparseMatrix<real_type> & J, RealVect & b, bool has_just_been_inialized){
     // solves (for x) the linear system J.x = b
     // supposes that the solver has been initialized (call klu_solver.analyze() before calling that)
     // J is const even if it does not compile if said const
@@ -88,11 +88,11 @@ int NICSLULinearSolver::solve(Eigen::SparseMatrix<real_type> & J, RealVect & b, 
         //        ret = klu_refactor(J_.outerIndexPtr(), J_.innerIndexPtr(), J_.valuePtr(), symbolic_, numeric_, &common_);
 
         // solver.FactorizeMatrix(ax, 0); //use all created threads
-        ret  = solver_.FactorizeMatrix(J.valuePtr(), nb_thread_);
+        ret  = solver_.FactorizeMatrix(J.valuePtr(), nb_thread_);  // TODO maybe 0 instead of nb_thread_ here, see https://github.com/chenxm1986/nicslu/blob/master/nicslu202110/demo/demo2.cpp
         if (ret < 0) {
-            std::cout << "NICSLULinearSolver::solve solver_.FactorizeMatrix error: " << ret << std::endl;
+            // std::cout << "NICSLULinearSolver::solve solver_.FactorizeMatrix error: " << ret << std::endl;
             // https://github.com/chenxm1986/nicslu/blob/master/nicslu202103/include/nicslu.h for error info
-            err = 2;
+            err = ErrorType::SolverReFactor;
             stop = true;
         }
     }
@@ -102,8 +102,8 @@ int NICSLULinearSolver::solve(Eigen::SparseMatrix<real_type> & J, RealVect & b, 
         x = RealVect(n);
         ret = solver_.Solve(&b(0), &x(0));
         if (ret < 0) {
-            std::cout << "NICSLULinearSolver::solve solver_.Solve error: " << ret << std::endl;
-            err = 3;
+            // std::cout << "NICSLULinearSolver::solve solver_.Solve error: " << ret << std::endl;
+            err = ErrorType::SolverSolve;
         }
         b = x;
     }