+++ src/SecurityAnalysis.cpp
@@ -155,6 +155,7 @@ void SecurityAnalysis::compute(const CplxVect & Vinit, int max_iter, real_type t
     Eigen::Index cont_id = 0;
     bool conv;
     CplxVect V;
+    // int contingency = 0;
     for(const auto & coeffs_modif: _li_coeffs){
         auto timer_modif_Ybus = CustTimer();
         bool invertible = remove_from_Ybus(Ybus, coeffs_modif);
@@ -166,15 +167,20 @@ void SecurityAnalysis::compute(const CplxVect & Vinit, int max_iter, real_type t
         // 5.2ms without it vs 81.9ms with it (for the iee 118)
         // So better make the computation, even if it's not used...
 
-        // if(invertible)
-        // {
-        V = Vinit_solver; // Vinit is reused for each contingencies
-        conv = compute_one_powerflow(Ybus, V, Sbus,
-                                     slack_ids, slack_weights,
-                                     bus_pv, bus_pq,
-                                     max_iter,
-                                     tol / sn_mva);
-        // }
+        if(invertible)
+        {
+            V = Vinit_solver; // Vinit is reused for each contingencies
+            conv = compute_one_powerflow(Ybus, V, Sbus,
+                                        slack_ids, slack_weights,
+                                        bus_pv, bus_pq,
+                                        max_iter,
+                                        tol / sn_mva);
+        }
+        // std::string conv_str =  conv ? "has converged" : "has diverged";
+        // std::cout << "contingency " << contingency << ": " << conv_str << std::endl;
+        // if(!conv) std::cout << "\t error was: " << _solver.get_error() << std::endl;
+        // ++contingency;
+
         timer_modif_Ybus = CustTimer();
         readd_to_Ybus(Ybus, coeffs_modif);
         _timer_modif_Ybus += timer_modif_Ybus.duration();